requirements.txt
ipykernel
legrad_torch
torch
Pillow
ftfy
regex
tqdm
open_clip_torch